refactored and cleanup code. Added Map Component. Added Tree and Layer Util files. some bug fixes
