+++ frontend/src/Components/LayerTree/LayerTree.tsx
@@ -1,218 +1,23 @@
+import { ExpandOutlined } from '@ant-design/icons/lib/icons';
+import { useMap } from '@terrestris/react-geo';
+import { Menu } from 'antd';
 import { Key } from 'antd/lib/table/interface';
-import Collection from 'ol/Collection';
 import BaseLayer from 'ol/layer/Base';
 import LayerGroup from 'ol/layer/Group';
 import ImageLayer from 'ol/layer/Image';
-import OlMap from 'ol/Map';
+import { transformExtent } from 'ol/proj';
 import ImageWMS from 'ol/source/ImageWMS';
 import { getUid } from 'ol/util';
-import React, { ReactNode, useEffect, useState } from 'react';
-import { JsonApiResponse } from '../../Repos/JsonApiRepo';
+import React, { useEffect, useState } from 'react';
 import LayerRepo from '../../Repos/LayerRepo';
-import { TreeFormField, TreeNodeType } from '../Shared/FormFields/TreeFormField/TreeFormField';
+import { LayerUtils } from '../../Utils/LayerUtils';
+import { TreeUtils } from '../../Utils/TreeUtils';
+import { TreeFormField } from '../Shared/FormFields/TreeFormField/TreeFormField';
+import { TreeNodeType } from '../Shared/FormFields/TreeFormField/TreeFormFieldTypes';
+import { LayerTreeProps } from './LayerTreeTypes';
 
-type OlWMSServerType = 'ESRI' | 'GEOSERVER' | 'MAPSERVER' | 'QGIS';
-
-export interface CreateLayerOpts {
-  url: string;
-  version: '1.1.0' | '1.1.1' | '1.3.0';
-  format: 'image/jpeg' | 'image/png';
-  layers: string;
-  visible: boolean;
-  serverType: OlWMSServerType;
-  mrMapLayerId?: string | number;
-  legendUrl: string;
-  title: string;
-  name?: string;
-  properties: Record<string, string>;
-  extent?: any[]
-}
-
-interface LayerTreeProps {
-  map: OlMap
-  layerGroup?: LayerGroup;
-  asyncTree?: boolean;
-  addLayerDispatchAction?:(
-    nodeAttributes: any,
-    newNodeParent?: string | number | null | undefined) =>
-    Promise<JsonApiResponse> | void;
-  removeLayerDispatchAction?: (nodeToRemove: TreeNodeType) => Promise<JsonApiResponse> | void;
-  editLayerDispatchAction?: (nodeId:number|string, nodeAttributesToUpdate: any) => Promise<JsonApiResponse> | void;
-  dragLayerDispatchAction?: (nodeBeingDraggedInfo: any) => Promise<JsonApiResponse> | void;
-  selectLayerDispatchAction?: (selectedKeys: Key[], info: any) => void;
-  layerAttributeForm?: ReactNode;
-}
-
-// export const setLayerSourceWMSParams = (layer: ImageLayer<ImageWMS>) => {
-//   const source = layer.getSource();
-//   source.setUrl('');
-//   source.serverType:
-// }
-
-export const getAllMapLayers = (collection: OlMap | LayerGroup): (LayerGroup | BaseLayer | ImageLayer<ImageWMS>)[] => {
-  if (!(collection instanceof OlMap) && !(collection instanceof LayerGroup)) {
-    console.error('Input parameter collection must be from type `ol.Map` or `ol.layer.Group`.');
-    return [];
-  }
-
-  const layers = collection.getLayers().getArray();
-  const allLayers:any = [];
-
-  layers.forEach((layer) => {
-    if (layer instanceof LayerGroup) {
-      getAllMapLayers(layer).forEach((layeri:any) => allLayers.push(layeri));
-    }
-    allLayers.push(layer);
-  });
-  
-  return allLayers;
-};
-
-export const createMrMapOlWMSLayer = (opts: CreateLayerOpts): ImageLayer<ImageWMS> => {
-  const olLayerSource = new ImageWMS({
-    url: opts.url,
-    params: {
-      'LAYERS': opts.layers,
-      'VERSION': opts.version,
-      'FORMAT': opts.format,
-      'TRANSPARENT': true
-    },
-    serverType: opts.serverType
-  });
-
-  const olWMSLayer = new ImageLayer({
-    source: olLayerSource,
-    visible: opts.visible,
-  });
-
-  olWMSLayer.setProperties({
-    mrMapLayerId: opts.mrMapLayerId,
-    legendUrl: opts.legendUrl,
-    title: opts.title,
-    name: opts.name,
-    extent: opts.extent,
-    ...opts.properties
-  });
-
-  return olWMSLayer;
-};
-
-export const getLayerByMrMapLayerId= (
-  collection: OlMap | LayerGroup, 
-  id: string | number
-): LayerGroup | BaseLayer | ImageLayer<ImageWMS> | undefined => {
-  const layersToSearch = getAllMapLayers(collection);
-  return layersToSearch.find((layer:any) => String(layer.getProperties().mrMapLayerId) === String(id));
-};
-
-export const getLayerGroupByName = (
-  collection: OlMap | LayerGroup, 
-  layerGroupName: string
-): LayerGroup | undefined => {
-  const requiredLayerGroup = getAllMapLayers(collection)
-    .find((layer:any) => layer.getProperties().name === layerGroupName);
-  if(requiredLayerGroup instanceof LayerGroup){
-    return requiredLayerGroup;
-  }
-};
-
-export const getLayerGroupByMrMapLayerId = (
-  collection: OlMap | LayerGroup, 
-  id: string|number
-): LayerGroup | undefined => {
-  const allMapLayers = getAllMapLayers(collection);
-  const requiredLayerGroup = allMapLayers
-    .find((layer:any) => layer.getProperties().mrMapLayerId === id);
-  if(requiredLayerGroup) {
-    if(requiredLayerGroup instanceof LayerGroup) {
-      return requiredLayerGroup;
-    } else {
-      if(requiredLayerGroup.getProperties().parent) {
-        const parentLayer = allMapLayers
-          .find((layer:any) => layer.getProperties().mrMapLayerId === requiredLayerGroup.getProperties().parent);
-        if(parentLayer && parentLayer instanceof LayerGroup) {
-          return parentLayer;
-        }
-      } else {
-        if (collection instanceof LayerGroup) {
-          return collection;
-        }
-        if (collection instanceof OlMap) {
-          collection.getLayerGroup();
-        }
-      }
-    }
-  } else {
-    if (collection instanceof LayerGroup) {
-      return collection;
-    }
-    if (collection instanceof OlMap) {
-      collection.getLayerGroup();
-    }
-  }
-};
-
-export const addLayerToGroupByName = (
-  collection: OlMap | LayerGroup, 
-  layerGroupName: string, 
-  layerToAdd: LayerGroup | BaseLayer
-): void => {
-  const layerGroup: LayerGroup | undefined = getLayerGroupByName(collection, layerGroupName);
-  if(layerGroup) {
-    const layerArray = layerGroup.getLayers().getArray();
-    layerArray.push(layerToAdd);
-    layerGroup.setLayers(new Collection(layerArray));
-  } else {
-    console.warn(`No layer group with the name ${layerGroupName}, was found on the map`);
-  }
-};
-
-export const addLayerToGroupByMrMapLayerId = (
-  collection: OlMap | LayerGroup, 
-  id: string | number, 
-  layerToAdd: LayerGroup | BaseLayer
-): void => {
-  const layerGroup: LayerGroup | undefined = getLayerGroupByMrMapLayerId(collection, id);
-  if(layerGroup) {
-    const layerArray = layerGroup.getLayers().getArray();
-    layerArray.push(layerToAdd);
-    layerGroup.setLayers(new Collection(layerArray));
-  } else {
-    console.warn(`No layer group with the id ${id}, was found on the map`);
-  }
-};
-
-export const OlLayerGroupToTreeNodeList = (layerGroup: LayerGroup): TreeNodeType[] => {
-  return layerGroup.getLayers().getArray().map((layer: LayerGroup | BaseLayer) => {
-    const node: any = {
-      key: layer.getProperties().mrMapLayerId,
-      title: layer.getProperties().title,
-      parent: layer.getProperties().parent,
-      properties: layer.getProperties(),
-      isLeaf: true,
-      expanded: layer instanceof LayerGroup,
-      children: []
-    }; 
-    if (layer instanceof LayerGroup ) {
-      node.children = OlLayerGroupToTreeNodeList(layer);
-      node.isLeaf = false;
-    }
-    return node;
-  });
-};
-
-export const OlLayerToTreeNode = (layer: BaseLayer): TreeNodeType => {
-  const node: any = {
-    key: layer.getProperties().mrMapLayerId,
-    title: layer.getProperties().title,
-    parent: layer.getProperties().parent,
-    properties: layer.getProperties(),
-    isLeaf: true,
-    expanded: layer instanceof LayerGroup,
-    children: []
-  }; 
-  return node;
-};
+const treeUtils =  new TreeUtils();
+const layerUtils =  new LayerUtils();
 
 const layerTreeLayerGroup = new LayerGroup({
   opacity: 1,
@@ -224,7 +29,7 @@ const layerTreeLayerGroup = new LayerGroup({
 });
 
 export const LayerTree = ({
-  map,
+  // map,
   layerGroup = layerTreeLayerGroup,
   asyncTree = false,
   addLayerDispatchAction = () => undefined,
@@ -238,12 +43,12 @@ export const LayerTree = ({
   // not to the tree form field component.
   // The tree form field component handles generic logic for a tree, not for the layers or interaction with map.
   // Only change it if you detect aa bug thaat could be traced baack deep to the tree form field
-
+  const map = useMap();
   const [treeData, setTreeData] = useState<TreeNodeType[]>([]);
   
   useEffect(() => {
     const onLayerGroupChange = (e:any) => {
-        setTreeData(OlLayerGroupToTreeNodeList(layerGroup));
+        setTreeData(treeUtils.OlLayerGroupToTreeNodeList(layerGroup));
     };
     const setWMSParams = async(theLayer: ImageLayer<ImageWMS>) => {
       try {
@@ -259,8 +64,8 @@ export const LayerTree = ({
       }
     };
     map.addLayer(layerGroup);
-    setTreeData(OlLayerGroupToTreeNodeList(layerGroup));
-    const allMapLayers = getAllMapLayers(layerGroup);
+    setTreeData(treeUtils.OlLayerGroupToTreeNodeList(layerGroup));
+    const allMapLayers = layerUtils.getAllMapLayers(layerGroup);
     allMapLayers.forEach((mapLayer: any) => {
       if(mapLayer instanceof ImageLayer) {
         const src: ImageWMS = mapLayer.getSource();
@@ -283,7 +88,7 @@ export const LayerTree = ({
   const onCheckLayer = (checkedKeys: (Key[] | {checked: Key[]; halfChecked: Key[];}), info: any) => {
     const { checked } = info;
     const eventKey = info.node.key;
-    const layer = getLayerByMrMapLayerId(map, eventKey);
+    const layer = layerUtils.getLayerByMrMapLayerId(map, eventKey);
     setLayerVisibility(layer, checked);
   };
 
@@ -326,10 +131,40 @@ export const LayerTree = ({
     });
   };
 
-  // const onAddNewLayer = (nodeAttributes: any, newNodeParent?: string | number | null | undefined, layer: any) =>  {
-  //   addLayerDispatchAction(nodeAttributes, newNodeParent);
-  //   addLayerToGroup(map, layerGroup.getProperties().name, layer);
-  // };
+  const onFitToLayerExtent = async(layerId: string) => {
+    try {
+      const res = await new LayerRepo().autocompleteInitialValue(layerId);
+      if(res.attributes.WMSParams.bbox) {
+        map.getView().fit(transformExtent(res.attributes.WMSParams.bbox, 'EPSG:4326', 'EPSG:3857'));
+      }
+    } catch(error) {
+      //@ts-ignore
+      throw new Error(error);
+    }
+  };
+
+  const layerActions = (nodeData: TreeNodeType|undefined): any => {
+    return (
+      <>
+      {/* <Divider orientation='right' plain /> */}
+      <Menu.Item
+        onClick={async() => {
+          // fit to layer extent
+          const theLayer = layerUtils.getAllMapLayers(layerGroup).find(l => l.get('key') === nodeData?.key);
+          if(theLayer && theLayer.get('renderingLayer')) {
+            onFitToLayerExtent(theLayer.get('renderingLayer'));
+          } else {
+            console.warn('Layer not found');
+          }
+        }}
+        icon={<ExpandOutlined/>}
+        key='zoom-to-extent'
+      >
+        Zoom to layer Extent
+      </Menu.Item>
+      </>
+    );
+  };
 
   return (
     <TreeFormField
@@ -347,6 +182,7 @@ export const LayerTree = ({
       attributeContainer='drawer'
       contextMenuOnNode
       checkableNodes
+      extendedNodeActions={layerActions}
     />
   );
 };
