+++ frontend/src/Components/TheMap/TheMap.css
@@ -1,0 +1,15 @@
+.Map {
+    position: relative;
+    height: 100%;
+    width: 100%;
+  }
+  
+.Map #the-map {
+    height: 100%;
+    width: 100%;
+}
+
+.ol-zoom {
+    top: unset;
+    bottom: 0.5em;
+  }
